automation/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.2
execnet==2.0.2
idna==3.4
iniconfig==2.0.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pluggy==1.2.0
pytest==7.3.1
pytest-rerunfailures==11.1.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.7
